Add reversFunction and test
