+++ package.json
@@ -15,5 +15,8 @@
   "bugs": {
     "url": "https://github.com/karam084/testing-practice/issues"
   },
-  "homepage": "https://github.com/karam084/testing-practice#readme"
+  "homepage": "https://github.com/karam084/testing-practice#readme",
+  "dependencies": {
+    "express": "^4.17.3"
+  }
 }
