+++ reverseFunc.test.js
@@ -1,0 +1,4 @@
+const reverseString = require('./reverseFunction');
+test('return reverse string', () => {
+  expect(reverseString('hello')).toBe('olleh');
+});
