+++ reverseFunction.js
@@ -1,0 +1,10 @@
+function reverseString(str) {
+  let newString = '';
+  for (let i = str.length - 1; i >= 0; i--) {
+    newString += str[i];
+  }
+  return newString;
+}
+reverseString('hello');
+
+module.exports = reverseString;
