+++ stringfunction.js
@@ -8,12 +8,3 @@ function end_script() {
   }
 }
 module.exports = end_script;
-
-function reverseString(str) {
-  const newString = '';
-  for (let i = str.length - 1; i >= 0; i--) {
-    newString += str[i];
-  }
-  return newString;
-}
-reverseString('hello');
